document-Q&A-gemma-groq/requirements.txt
faiss-cpu
groq
langchain_groq
PyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pypdf 
